+++ Network/Telegram/API/Bot/Capacity.hs
@@ -3,3 +3,4 @@ module Network.Telegram.API.Bot.Capacity (module Exports) where
 import Network.Telegram.API.Bot.Capacity.Purgeable as Exports
 import Network.Telegram.API.Bot.Capacity.Postable as Exports
 import Network.Telegram.API.Bot.Capacity.Editable as Exports
+import Network.Telegram.API.Bot.Capacity.Droppable as Exports