+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # 0.1.1
 * Define `chat` lens for `Update` datatype
+* Add string identificator to `Callback` datatype
 * Define `Notification` datatype
+* Define `Droppable` typeclass